Advances on code for potential fields

+++ MATLABCODE/Potentialfields/uavgenerator3d.m
@@ -1,15 +1,15 @@
 function [totaldrones, inidrones, goaldrones, colordrones] = uavgenerator3d (min, max, raddrones)
     %Generate UAVs
-    totaldrones = randi([min,max],1,1);
-    inidrones = zeros(totaldrones,3);
-    goaldrones = zeros(totaldrones,3);
-    colordrones = zeros(totaldrones,3);
+    totaldrones = single(randi([min,max],1,1));
+    inidrones = single(zeros(totaldrones,3));
+    goaldrones = single(zeros(totaldrones,3));
+    colordrones = single(zeros(totaldrones,3));
     for i=1:totaldrones
-         inidrones(i,:) = rand(1,3) * 30;
-         goaldrones(i,:) = rand(1,3) * 30;
-         colordrones(i,1) = rand;
-         colordrones(i,2) = rand;
-         colordrones(i,3) = rand;
+         inidrones(i,:) = single(rand(1,3) * 30);
+         goaldrones(i,:) =  single(rand(1,3) * 30);
+         colordrones(i,1) = single(rand);
+         colordrones(i,2) = single(rand);
+         colordrones(i,3) = single(rand);
     end
     
     %Print initial positions and goals
